vanilla_etl/requirements.txt
accelerate==0.34.0
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
apprise==1.8.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgi-lifespan==2.1.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.8.0
bleach==6.1.0
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cookiecutter==2.6.0
cookiecutter-data-science==2.0.0
coolname==2.2.0
croniter==2.0.7
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.8
dask==2024.8.2
dateparser==1.2.0
debugpy==1.8.5
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
distributed==2024.8.2
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.15.4
filetype==1.2.0
flake8==7.1.1
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
ghp-import==2.1.0
google-auth==2.34.0
graphviz==0.20.3
greenlet==3.0.3
griffe==0.47.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
humanize==4.10.0
hyperframe==5.2.0
idna==2.10
imbalanced-learn==0.12.3
imblearn==0.0
importlib_metadata==8.4.0
importlib_resources==6.1.3
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jinja2-humanize-extension==0.4.0
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
kubernetes==29.0.0
langcodes==3.4.0
language_data==1.2.0
locket==1.0.0
Mako==1.3.5
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.0.2
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.43.0
openpyxl==3.1.5
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pendulum==2.1.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prefect==2.20.0
prefect-dask==0.2.10
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.4
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
pytzdata==2020.1
PyYAML==6.0.2
pyyaml_env_tag==0.1
pyzmq==26.2.0
readchar==4.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.2
tabula-py==2.9.3
tabulate==0.9.0
tblib==3.0.0
terminado==0.18.1
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
trio==0.26.2
typer==0.12.5
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.6
-e git+https://github.com/abad623/vanilla_steel.git@ddc7fdbe8dc71ee1a2e9e80c761fac5d0e08ca7b#egg=vanilla_etl
wasabi==1.1.3
watchdog==5.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
zict==3.0.0
zipp==3.20.1
